requirements.txt
Flask
nltk
transformers
torch
psycopg2-binary
